Changed routes so they are updated according to the new controllers location

+++ routes/web.php
@@ -20,22 +20,22 @@ Route::get('/', 'App\Http\Controllers\HomeController@index')->name('home.index')
 Route::middleware(['auth.role:admin'])->group(function () {
     //Put all admin routes with prefix /admin
     Route::get('/admin', 'App\Http\Controllers\Admin\AdminController@index')->name('admin.index');
-    Route::get('/admin/part/', 'App\Http\Controllers\PartController@showAll')->name('admin.part.showall');
-    Route::get('/admin/part/create', 'App\Http\Controllers\PartController@create')->name('admin.part.create');
-    Route::post('/admin/part/save', 'App\Http\Controllers\PartController@save')->name('admin.part.save');
-    Route::get('/admin/part/show/{id}', 'App\Http\Controllers\PartController@show')->name('admin.part.show');
-    Route::delete('/admin/part/remove/{id}', 'App\Http\Controllers\PartController@remove')->name('admin.part.remove');
+    Route::get('/admin/part/', 'App\Http\Controllers\Admin\PartController@showAll')->name('admin.part.showall');
+    Route::get('/admin/part/create', 'App\Http\Controllers\Admin\PartController@create')->name('admin.part.create');
+    Route::post('/admin/part/save', 'App\Http\Controllers\Admin\PartController@save')->name('admin.part.save');
+    Route::get('/admin/part/show/{id}', 'App\Http\Controllers\Admin\PartController@show')->name('admin.part.show');
+    Route::delete('/admin/part/remove/{id}', 'App\Http\Controllers\Admin\PartController@remove')->name('admin.part.remove');
 });
 
 //User routes
 Route::middleware(['auth.role:user'])->group(function () {
     //Put all user routes with prefix /user
     Route::get('/user', 'App\Http\Controllers\User\UserController@index')->name('user.index');
-    Route::get('/user/showAll', 'App\Http\Controllers\BikeController@showAll')->name("user.bike.showAll");
-    Route::get('/user/create', 'App\Http\Controllers\BikeController@create')->name("user.bike.create");
-    Route::post('/user/save', 'App\Http\Controllers\BikeController@save')->name("user.bike.save");
-    Route::get('/user/show/{id}', 'App\Http\Controllers\BikeController@show')->name("user.bike.show");
-    Route::get('/remove/{id}', 'App\Http\Controllers\BikeController@remove')->name("user.bike.remove");
+    Route::get('/user/showAll', 'App\Http\Controllers\User\BikeController@showAll')->name("user.bike.showAll");
+    Route::get('/user/create', 'App\Http\Controllers\User\BikeController@create')->name("user.bike.create");
+    Route::post('/user/save', 'App\Http\Controllers\User\BikeController@save')->name("user.bike.save");
+    Route::get('/user/show/{id}', 'App\Http\Controllers\User\BikeController@show')->name("user.bike.show");
+    Route::delete('/remove/{id}', 'App\Http\Controllers\User\BikeController@remove')->name("user.bike.remove");
 });
 
 //Auth routes
